Move UuidRepresentationOverridingCodecRegistry to org.bson.internal
