+++ bson/src/main/org/bson/internal/ChildCodecRegistry.java
@@ -14,10 +14,11 @@
  * limitations under the License.
  */
 
-package org.bson.codecs.configuration;
+package org.bson.internal;
 
 
 import org.bson.codecs.Codec;
+import org.bson.codecs.configuration.CodecRegistry;
 
 /**
  * An implementation of CodecRegistry that is used to detect cyclic dependencies between Codecs
@@ -25,13 +26,13 @@ import org.bson.codecs.Codec;
  * @param <T> the {@code Codec} type
  * @since 3.12
  */
-public class ChildCodecRegistry<T> implements CodecRegistry {
+class ChildCodecRegistry<T> implements CodecRegistry {
 
     private final ChildCodecRegistry<?> parent;
     private final CycleDetectingCodecRegistry registry;
     private final Class<T> codecClass;
 
-    public ChildCodecRegistry(final CycleDetectingCodecRegistry registry, final Class<T> codecClass) {
+    ChildCodecRegistry(final CycleDetectingCodecRegistry registry, final Class<T> codecClass) {
         this.codecClass = codecClass;
         this.parent = null;
         this.registry = registry;
