+++ bson/src/main/org/bson/internal/LazyCodec.java
@@ -14,13 +14,14 @@
  * limitations under the License.
  */
 
-package org.bson.codecs.configuration;
+package org.bson.internal;
 
 import org.bson.BsonReader;
 import org.bson.BsonWriter;
 import org.bson.codecs.Codec;
 import org.bson.codecs.DecoderContext;
 import org.bson.codecs.EncoderContext;
+import org.bson.codecs.configuration.CodecRegistry;
 
 class LazyCodec<T> implements Codec<T> {
     private final CodecRegistry registry;
